# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Test API"
        }
    ],
    "paths": {
...
        "/dashboard/count": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get list count in dashboard",
                "description": "Get list count in dashboard",
                "operationId": "showDashboardCount",
                "parameters": [
                    {
                        "name": "d",
                        "in": "query",
                        "description": "search by day",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/dashboard/top-user-orders": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "top user orders",
                "description": "Get top user orders",
                "operationId": "showDashboardUserOrder",
                "parameters": [
                    {
                        "name": "d",
                        "in": "query",
                        "description": "search by day",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/dashboard/top-sellers": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "get top sellers",
                "description": "get top sellers",
                "operationId": "showDashboardSellers",
                "parameters": [
                    {
                        "name": "d",
                        "in": "query",
                        "description": "search by day",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}